# openapi.yaml
openapi: 3.0.0
info:
  version: '1.0'
  title: Lbaw Project - MyRel
  description: Web Resources Specification (A7) for Lbaw Project
servers:
  - url: http://lbaw2212.lbaw.fe.up.pt
...
paths:
  /login:
    get:
      operationId: R101
      summary: "R101: Login Form"
      description: "Provide login form(UI12). Access: PUB"
      tags:
        - "M01: Authentication"
      responses:
        "200":
          description: "Return login page."
    post:
      operationId: R102
      summary: "R102: Login Action"
      description: "Processes the login form data. Access: PUB"
      tags:
        - "M01: Authentication"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                  format: password
              required:
                - email
                - password
      responses:
        "302":
          description: "Redirect after processing the login credentials."
          headers:
            location:
              schema:
                type: string
              examples:
                302Success:
                  description: "Login succeed. Will redirect to user feed."
                  value: "/posts"
        "401":
          description: "Redirect after processing the login credentials."
          headers:
            location:
              schema:
                type: string
              examples:
                401Error:
                  description: "Unauthorized. Response will be processed in the login page to alert the user about the failure."
                  value: "/login"
...
  /api/posts:
    put:
      operationId: R402
      summary: "R402: Create a Post"
      description: "Create a new post. Access: USR, ADM"
      tags:
        - "M04: Posts"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                text:
                  type: string
                visibility:
                  type: string
                photo:
                  type: string
                  format: binary
              required:
                - text
                - visibility
      responses:
        "302":
          description: "Redirect after creating new post."
          headers:
            location:
              schema:
                type: string
              examples:
                302Success:
                  description: "Successful create post. Redirecting to post page"
                  value: "/post/{id}"
                302Error:
                  description: "Unsuccessful create post. Redirecting back to posts page"
                  value: "/posts"
...